+++ implementations/java/docker/Dockerfile-driver-did-stack
@@ -16,7 +16,7 @@ ENV PATH $JAVA_HOME/bin:$PATH
 
 RUN apt-get install -y --no-install-recommends git
 
-# build driver-did-bsk
+# build driver-did-stack
 
 ADD . /opt/uni-resolver-java
 
@@ -24,7 +24,7 @@ RUN cd /opt/uni-resolver-java && mvn clean
 RUN cd /opt/uni-resolver-java && mvn install -N -DskipTests
 RUN cd /opt/uni-resolver-java/uni-resolver-core && mvn install -N -DskipTests
 RUN cd /opt/uni-resolver-java/driver && mvn install -N -DskipTests
-RUN cd /opt/uni-resolver-java/driver-did-bsk && mvn install package -N -DskipTests
+RUN cd /opt/uni-resolver-java/driver-did-stack && mvn install package -N -DskipTests
 
 # done
 
@@ -32,4 +32,4 @@ RUN apt-get clean && rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/*
 
 EXPOSE 8084
 
-CMD ["/opt/uni-resolver-java/docker/run-driver-did-bsk.sh"]
+CMD ["/opt/uni-resolver-java/docker/run-driver-did-stack.sh"]
