--- implementations/java/docker/run-driver-did-bsk.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-cd /opt/uni-resolver-java/driver-did-bsk/
-mvn jetty:run -P war
