+++ implementations/java/docker/run-driver-did-stack.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+cd /opt/uni-resolver-java/driver-did-stack/
+mvn jetty:run -P war
