+++ implementations/java/driver-did-stack/pom.xml
@@ -2,9 +2,9 @@
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
 	<modelVersion>4.0.0</modelVersion>
-	<artifactId>uni-resolver-driver-did-bsk</artifactId>
+	<artifactId>uni-resolver-driver-did-stack</artifactId>
 	<packaging>${packaging.type}</packaging>
-	<name>uni-resolver-driver-did-bsk</name>
+	<name>uni-resolver-driver-did-stack</name>
 
 	<parent>
 		<groupId>decentralized-identity</groupId>
