+++ implementations/java/driver-did-stack/src/main/webapp/WEB-INF/applicationContext.xml
@@ -7,7 +7,7 @@
 
 	<bean name="DriverServlet" class="uniresolver.driver.DriverServlet">
 		<property name="driver">
-           <bean id="DidBskDriver" class="uniresolver.driver.did.bsk.DidBskDriver">
+           <bean id="DidStackDriver" class="uniresolver.driver.did.stack.DidStackDriver">
 			</bean>
 		</property>
 	</bean>
