+++ implementations/java/driver-did-stack/src/test/resources/log4j.properties
