+++ implementations/java/examples/pom.xml
@@ -39,7 +39,7 @@
         </dependency>		
         <dependency>
 			<groupId>decentralized-identity</groupId>
-			<artifactId>uni-resolver-driver-did-bsk</artifactId>
+			<artifactId>uni-resolver-driver-did-stack</artifactId>
 			<version>0.1-SNAPSHOT</version>
 			<scope>compile</scope>
 		</dependency>
