+++ implementations/java/examples/src/main/java/uniresolver/examples/TestLocalUniResolver.java
@@ -3,7 +3,7 @@ import uniresolver.ddo.DDO;
 import uniresolver.driver.did.btcr.DidBtcrDriver;
 import uniresolver.driver.did.btcr.bitcoinconnection.BlockcypherAPIExtendedBitcoinConnection;
 import uniresolver.driver.did.sov.DidSovDriver;
-import uniresolver.driver.did.bsk.DidBskDriver;
+import uniresolver.driver.did.stack.DidStackDriver;
 import uniresolver.local.LocalUniResolver;
 
 public class TestLocalUniResolver {
@@ -22,7 +22,7 @@ public class TestLocalUniResolver {
 		DDO ddo2 = uniResolver.resolve("did:btcr:xkrn-xzcr-qqlv-j6sl");
 		System.out.println(ddo2.serialize());
 
-        DDO ddo3 = uniResolver.resolve("did:bsk:v0:16EMaNw3pkn3v6f2BgnSSs53zAKH4Q8YJg-0");
+        DDO ddo3 = uniResolver.resolve("did:stack:v0:16EMaNw3pkn3v6f2BgnSSs53zAKH4Q8YJg-0");
         System.out.println(ddo3.serialize());
 	}
 }
