+++ implementations/java/pom.xml
@@ -25,7 +25,7 @@
 		<module>uni-resolver-client</module>
 		<module>uni-resolver-web</module>
 		<module>driver</module>
-		<module>driver-did-bsk</module>
+		<module>driver-did-stack</module>
 		<module>driver-did-btcr</module>
 		<module>driver-did-sov</module>
 		<module>driver-http</module>
