+++ implementations/java/uni-resolver-local/pom.xml
@@ -51,7 +51,7 @@
                  </dependency>
                  <dependency>
                     <groupId>decentralized-identity</groupId>
-                    <artifactId>uni-resolver-driver-did-bsk</artifactId>
+                    <artifactId>uni-resolver-driver-did-stack</artifactId>
                     <version>0.1-SNAPSHOT</version>
                     <scope>compile</scope>
                  </dependency>
