+++ implementations/java/uni-resolver-web/pom.xml
@@ -109,7 +109,7 @@
 				</dependency>
 				<dependency>
 					<groupId>decentralized-identity</groupId>
-					<artifactId>uni-resolver-driver-did-bsk</artifactId>
+					<artifactId>uni-resolver-driver-did-stack</artifactId>
 					<version>0.1-SNAPSHOT</version>
 					<scope>compile</scope>
 				</dependency>
