+++ implementations/java/uni-resolver-web/src/main/webapp/WEB-INF/applicationContext.xml
@@ -15,7 +15,7 @@
 						<ref bean="DidUportDriver" />
 						<ref bean="DidV1Driver" />
                         <ref bean="DidIpidDriver" />
-                        <ref bean="DidBskDriver" />
+                        <ref bean="DidStackDriver" />
 					</util:list>
 				</property>
 			</bean>
@@ -71,9 +71,9 @@
 		<property name="rawDidIdentifier" value="true" />
      </bean>
 
-     <bean id="DidBskDriver" class="uniresolver.driver.http.HttpDriver">
-        <property name="driverUri" value="http://driver-did-bsk:8084/1.0/dids/" />
-        <property name="pattern" value="^did:bsk:.+$" />
+     <bean id="DidStackDriver" class="uniresolver.driver.http.HttpDriver">
+        <property name="driverUri" value="http://driver-did-stack:8084/1.0/dids/" />
+        <property name="pattern" value="^did:stack:.+$" />
      </bean>
 
 </beans>
